Move pod related functions file in test/e2e/framework/util.go to its sub direcotry
