+++ test/e2e/framework/kubelet/BUILD
@@ -2,9 +2,18 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["stats.go"],
+    srcs = [
+        "kubelet_pods.go",
+        "stats.go",
+    ],
     importpath = "k8s.io/kubernetes/test/e2e/framework/kubelet",
     visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/master/ports:go_default_library",
+        "//staging/src/k8s.io/api/core/v1:go_default_library",
+        "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
+        "//test/e2e/framework/node:go_default_library",
+    ],
 )
 
 filegroup(
